app/requirements.txt
#Django==4.0.2
Django==3.2.25
asgiref==3.5.0
sqlparse==0.4.4
debugpy==1.5.1
mysqlclient==2.1.0
gunicorn==20.1.0

certifi==2023.7.22
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==42.0.4
defusedxml==0.7.1
django-allauth==0.49.0
idna==3.7
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.4.0
pytest==6.2.5
pytest-django==4.5.2
python3-openid==3.2.0
requests==2.31.0
requests-cache==1.0.1
requests-oauthlib==1.3.1
urllib3==1.26.18
requests-mock==1.11.0

#OLD
#Django==3.1.14
#django-allauth==0.44.0
django-crispy-forms==1.9.2
django-debug-toolbar>=3.2.4
django-extensions>=3.0.9
django-filter>=2.4.0
django-redis==5.2.0
django-select2~=7.2.0
django-simple-history~=3.0.0
django-userforeignkey~=0.3.0
djangorestframework~=3.11.0
#django-ratelimit>=3.0.1
Markdown~=3.0.1
#mysqlclient~=2.0.1
numpy>=1.15.4
packaging>=20.4
pandas>=1.3.0
#pip>=20.2.4
plotly>=4.12.0
#Old requirement, switch back if broken --> pytz~=2018.9
pytz>=2021.3
#requests~=2.23.0

#Celery
celery==5.2.7
celery-progress==0.1.2
django-celery-results==2.4.0
watchfiles==0.16.1
importlib-metadata<5.0

# pylint
pylint~=3.1.0
pylint-django~=2.5.5
